+++ src/main/java/it/gov/pagopa/fdr/repository/PaymentFullViewRepository.java
@@ -70,14 +70,6 @@ public class PaymentFullViewRepository extends Repository implements PanacheRepo
     return getPagedResult(resultPage);
   }
 
-  public PanacheQuery<PaymentFullViewEntity> findPageByFlowId(Long flowId, int pageNumber, int pageSize) {
-
-    Page page = Page.of(pageNumber, pageSize);
-    Sort sort = getSort(SortField.of(INDEX, Direction.Ascending));
-
-    return find(QUERY_GET_BY_FLOW_ID, sort, flowId).page(page);
-  }
-
   public RepositoryPagedResult<PaymentFullViewEntity> findByFlowId(
       Long flowId, int pageNumber, int pageSize) {
