+++ src/main/java/it/gov/pagopa/fdr/repository/entity/PaymentFullViewEntity.java
@@ -2,11 +2,7 @@ package it.gov.pagopa.fdr.repository.entity;
 
 import io.quarkus.hibernate.orm.panache.PanacheQuery;
 import io.quarkus.panache.common.Parameters;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import org.hibernate.annotations.Immutable;
@@ -17,6 +13,9 @@ import org.hibernate.annotations.Immutable;
 @Table(name = "payment_full_view")
 @Immutable
 public class PaymentFullViewEntity extends AbstractPaymentEntity {
+
+  @EmbeddedId
+  private PaymentId id;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "flow_id", referencedColumnName = "id", insertable = false, updatable = false)
