+++ src/main/java/it/gov/pagopa/fdr/repository/entity/PaymentStagingEntity.java
@@ -1,14 +1,9 @@
 package it.gov.pagopa.fdr.repository.entity;
 
-import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
-import io.quarkus.hibernate.orm.panache.PanacheQuery;
-import io.quarkus.panache.common.Parameters;
 import jakarta.persistence.*;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
-import java.math.BigDecimal;
-import java.time.Instant;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
@@ -16,6 +11,6 @@ import java.time.Instant;
 @Table(name = "payment_staging")
 public class PaymentStagingEntity extends AbstractPaymentEntity {
 
-  @Column(name = "org_id", length = 15)
-  public String orgId;
+  @EmbeddedId
+  private PaymentStagingId id;
 }
