+++ src/main/java/it/gov/pagopa/fdr/repository/entity/PaymentStagingId.java
@@ -1,0 +1,40 @@
+package it.gov.pagopa.fdr.repository.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+import java.util.Objects;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Embeddable
+public class PaymentStagingId implements Serializable {
+
+  @Column(name = "flow_id")
+  private Long flowId;
+
+  @Column(name = "index")
+  private Long index;
+
+  @Column(name = "org_id", length = 15)
+  public String orgId;
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) return true;
+    if (o == null || getClass() != o.getClass()) return false;
+    PaymentStagingId paymentId = (PaymentStagingId) o;
+    return Objects.equals(flowId, paymentId.flowId) && Objects.equals(index, paymentId.index) && Objects.equals(orgId, paymentId.orgId);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(flowId, index, orgId);
+  }
+}
+
