+++ src/main/java/it/gov/pagopa/fdr/schedule/HistoryCron.java
@@ -6,12 +6,10 @@ import io.quarkus.runtime.configuration.DurationConverter;
 import io.quarkus.scheduler.Scheduled;
 import it.gov.pagopa.fdr.repository.FlowRepository;
 import it.gov.pagopa.fdr.repository.FlowToHistoryRepository;
-import it.gov.pagopa.fdr.repository.PaymentFullViewRepository;
 import it.gov.pagopa.fdr.repository.PaymentRepository;
 import it.gov.pagopa.fdr.repository.entity.FlowEntity;
 import it.gov.pagopa.fdr.repository.entity.FlowToHistoryEntity;
 import it.gov.pagopa.fdr.repository.entity.PaymentEntity;
-import it.gov.pagopa.fdr.repository.entity.PaymentFullViewEntity;
 import it.gov.pagopa.fdr.storage.HistoryBlobStorageService;
 import it.gov.pagopa.fdr.storage.middleware.FlowBlobMapper;
 import it.gov.pagopa.fdr.storage.model.FlowBlob;
@@ -38,7 +36,7 @@ public class HistoryCron {
   private final HistoryBlobStorageService historyBlobStorageService;
   private final FlowToHistoryRepository flowToHistoryRepository;
   private final FlowRepository flowRepository;
-  private final PaymentFullViewRepository paymentFullViewRepository;
+  private final PaymentRepository paymentRepository;
   private final FlowBlobMapper flowBlobMapper;
 
   @ConfigProperty(name = "schedule.history.size")
@@ -58,14 +56,15 @@ public class HistoryCron {
           Logger log,
           FlowToHistoryRepository flowToHistoryRepository,
           FlowRepository flowRepository,
-          HistoryBlobStorageService historyBlobStorageService, PaymentFullViewRepository paymentFullViewRepository,
+          HistoryBlobStorageService historyBlobStorageService,
+          PaymentRepository paymentRepository,
           FlowBlobMapper flowBlobMapper) {
 
     this.log = log;
     this.flowToHistoryRepository = flowToHistoryRepository;
     this.flowRepository = flowRepository;
     this.historyBlobStorageService = historyBlobStorageService;
-    this.paymentFullViewRepository = paymentFullViewRepository;
+    this.paymentRepository = paymentRepository;
     this.flowBlobMapper = flowBlobMapper;
   }
 
@@ -183,9 +182,9 @@ public class HistoryCron {
   private List<PaymentBlob> handlePage(Long flowId) {
     int page = 0;
     List<PaymentBlob> result = new ArrayList<>();
-    PanacheQuery<PaymentFullViewEntity> payments;
+    PanacheQuery<PaymentEntity> payments;
     do {
-      payments = paymentFullViewRepository.findPageByFlowId(flowId, page, paymentPageSize);
+      payments = paymentRepository.findPageByFlowId(flowId, page, paymentPageSize);
       List<PaymentBlob> paymentsBlob =
           payments.stream().map(flowBlobMapper::toPaymentBlob).toList();
       result.addAll(paymentsBlob);
