+++ src/test/java/it/gov/pagopa/fdr/repository/PaymentStagingRepositoryIT.java
@@ -1,0 +1,6 @@
+package it.gov.pagopa.fdr.repository;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+class PaymentStagingRepositoryIT extends PaymentStagingRepositoryTest {}
